+++ android/src/main/kotlin/im/nfc/flutter_nfc_kit/FlutterNfcKitPlugin.kt
@@ -71,6 +71,13 @@ class FlutterNfcKitPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             return transceiveMethod.invoke(this, data) as ByteArray
         }
 
+        private fun ensureNfcHandler() {
+            if (!::nfcHandlerThread.isInitialized || !nfcHandlerThread.isAlive) {
+                nfcHandlerThread = HandlerThread("FlutterNfcKit-NfcHandlerThread").apply { start() }
+                nfcHandler = Handler(nfcHandlerThread.looper)
+            }
+        }
+
         private fun runOnNfcThread(result: Result, desc: String, fn: () -> Unit) {
             val handledFn = Runnable {
                 try {
@@ -89,13 +96,8 @@ class FlutterNfcKitPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
                     }
                 }
             }
-            val looperThread = nfcHandler.looper?.thread
-            if (looperThread == null || !looperThread.isAlive) {
-                val thread = HandlerThread("FlutterNfcKit").apply { start() }
-                nfcHandler = Handler(thread.looper)
-            }
-            val posted = nfcHandler.post(handledFn)
-            if (!posted) {
+            ensureNfcHandler()
+            if (!nfcHandler.post(handledFn)) {
                 result.error("500", "Failed to post job to NFC Handler thread.", null)
             }
         }
@@ -249,9 +251,7 @@ class FlutterNfcKitPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
     }
 
     override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
-        nfcHandlerThread = HandlerThread("NfcHandlerThread")
-        nfcHandlerThread.start()
-        nfcHandler = Handler(nfcHandlerThread.looper)
+        ensureNfcHandler()
 
         methodChannel = MethodChannel(flutterPluginBinding.binaryMessenger, "flutter_nfc_kit/method")
         methodChannel.setMethodCallHandler(this)